Chef relation is functional, all api routes are working
